requirements.txt
GDAL==3.11.0
GDAL==3.8.4
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
Pillow==12.1.1
PyYAML==6.0.3
rasterio==1.4.3
scipy==1.17.1
torch==2.11.0+cu130
torchvision==0.26.0+cu130
tqdm==4.67.1
transformers==4.57.3
ultralytics_thop==2.0.18
